# swagger.yaml
openapi: 3.0.0
info:
  title: Foot City
  contact: {}
  version: '1.0'
servers:
- url: http://example.com
  variables: {}
paths:
...
  /order:
    post:
      tags:
      - Order
      summary: Check Out
      description: This request checks out user cart list and adds them to order.
      operationId: CheckOut
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                total: 5000
                subtotal: 6000
                shipping: 1000
                orderStatus: pending
                addressDetails: []
                cart:
                - name: Nike
                  image: ./images/image.jpg
                  quantity: 3
                  size: xl
                  color: red
            example:
              total: 5000
              subtotal: 6000
              shipping: 1000
              orderStatus: pending
              addressDetails: []
              cart:
              - name: Nike
                image: ./images/image.jpg
                quantity: 3
                size: xl
                color: red
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  order:
                    subtotal: 6000
                    shipping: 1000
                    total: 5000
                    addressDetails: []
                    cart:
                    - name: Nike
                      image: ./images/image.jpg
                      quantity: 3
                      size: xl
                      color: red
                      _id: 63363b318cb7a4d511d7f66d
                    orderStatus: pending
                    user: 6329108dddd6b7fc854ccd02
                    _id: 63363b318cb7a4d511d7f66c
                    __v: 0
              example:
                order:
                  subtotal: 6000
                  shipping: 1000
                  total: 5000
                  addressDetails: []
                  cart:
                  - name: Nike
                    image: ./images/image.jpg
                    quantity: 3
                    size: xl
                    color: red
                    _id: 63363b318cb7a4d511d7f66d
                  orderStatus: pending
                  user: 6329108dddd6b7fc854ccd02
                  _id: 63363b318cb7a4d511d7f66c
                  __v: 0
      deprecated: false
    patch:
      tags:
      - Order
      summary: Update Cart
      description: This request updates the cart of an existing order
      operationId: UpdateCart
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                name: Nike
                image: ./images/image.jpg
                quantity: 3
                size: xl
                color: white
            example:
              name: Nike
              image: ./images/image.jpg
              quantity: 3
              size: xl
              color: white
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  order:
                    _id: 63363b318cb7a4d511d7f66c
                    subtotal: 6000
                    shipping: 1000
                    total: 5000
                    addressDetails: []
                    cart:
                    - name: Nike
                      image: ./images/image.jpg
                      quantity: 3
                      size: xl
                      color: white
                      _id: 63363bd48cb7a4d511d7f671
                    orderStatus: pending
                    user: 6329108dddd6b7fc854ccd02
                    __v: 0
              example:
                order:
                  _id: 63363b318cb7a4d511d7f66c
                  subtotal: 6000
                  shipping: 1000
                  total: 5000
                  addressDetails: []
                  cart:
                  - name: Nike
                    image: ./images/image.jpg
                    quantity: 3
                    size: xl
                    color: white
                    _id: 63363bd48cb7a4d511d7f671
                  orderStatus: pending
                  user: 6329108dddd6b7fc854ccd02
                  __v: 0
      deprecated: false
  /order/addAddressDetails/{order_id}:
    patch:
      tags:
      - Order
      summary: Add Address Details
      description: This request adds address details to the addressDetails field array in the order object
      operationId: AddAddressDetails
      parameters:
      - name: order_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                firstName: Maxwell
                lastName: Etombi
                email: etombimaxwell@gmail.com
                deliveryAddress: Plot 190, Extension 2
                city: Abuja
                region: Kubwa
            example:
              firstName: Maxwell
              lastName: Etombi
              email: etombimaxwell@gmail.com
              deliveryAddress: Plot 190, Extension 2
              city: Abuja
              region: Kubwa
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  order:
                    acknowledged: true
                    modifiedCount: 1
                    upsertedId: null
                    upsertedCount: 0
                    matchedCount: 1
              example:
                order:
                  acknowledged: true
                  modifiedCount: 1
                  upsertedId: null
                  upsertedCount: 0
                  matchedCount: 1
      deprecated: false
  /order/updateAddressDetails/{order_id}:
    patch:
      tags:
      - Order
      summary: Update Address Details
      description: This request updates a specific address info based on the addressDetails id
      operationId: UpdateAddressDetails
      parameters:
      - name: order_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: The updateId is the id of the addressDetails that wants to be updated. Hence, must be provided
        content:
          text/plain:
            schema:
              type: object
              description: The updateId is the id of the addressDetails that wants to be updated. Hence, must be provided
              example:
                updateId: 63363c8f8cb7a4d511d7f679
                firstName: MiriamElkanah
                lastName: Etombi
                email: etombimaxwell@gmail.com
                deliveryAddress: Plot 190, Extension 2
                city: Abuja
                region: Kubwa
            example:
              updateId: 63363c8f8cb7a4d511d7f679
              firstName: MiriamElkanah
              lastName: Etombi
              email: etombimaxwell@gmail.com
              deliveryAddress: Plot 190, Extension 2
              city: Abuja
              region: Kubwa
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: Address details updated successfully!
              example:
                msg: Address details updated successfully!
      deprecated: false
  /order/deleteAddressDetails/{order_id}:
    patch:
      tags:
      - Order
      summary: Delete Address Details
      description: This request deletes addressDetails based on the order id
      operationId: DeleteAddressDetails
      parameters:
      - name: order_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: The deleteId is the id of the specific address detail that wants to be removed from the list.
        content:
          text/plain:
            schema:
              type: object
              description: The deleteId is the id of the specific address detail that wants to be removed from the list.
              example:
                deleteId: 63363c8f8cb7a4d511d7f679
            example:
              deleteId: 63363c8f8cb7a4d511d7f679
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: Address details deleted successfully!
              example:
                msg: Address details deleted successfully!
      deprecated: false
...
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...